feat(workspaces): Spec 089 P5 — sharing and quota, and the two holes between correct features
Read/write is a platform concept. An earlier draft said whether a recipient may
commit is TermsJson, the product's business, reasoning that a platform enforcing
an "editor" role would be interpreting terms — which Spec 086 §6.1 forbids. That
is a bypassable authorisation boundary: the commit endpoint is a platform HTTP
endpoint, and a read-only recipient needs only to call it directly, with curl or
a modified client, which for an MIT-licensed product is a five-minute exercise.
The product-side check is decoration, and the grant reader's answer to "is there
a grant?" for a read-only recipient is yes.

So CircleGrant carries AccessLevel — three values, closed, no DSL — and
IShareGrantReader answers what a grant PERMITS rather than whether one exists.
TermsJson stays opaque and the product stays its only reader; what the platform
now knows is whether this caller may write to this container, which is the
access-control question every endpoint owner has to answer for itself. Unknown
levels resolve to None, so a future value cannot widen access by accident, and
two grants naming one resource take the highest — a stale read-only grant must
not quietly revoke write access the owner deliberately gave.

Stored bytes are a CEILING and emphatically not a counter. An earlier draft made
workspace-bytes a counter "drawn down as blobs are stored, returned on sweep",
and the second half is not implementable: CommitAsync moves the reservation into
Consumed, and Consumed never decreases — correctly, because a counter measures a
flow that cannot un-happen. A customer uploading 40GB and deleting it would have
permanently burned 40GB of a 50GB allowance while using none of it, and the
support ticket would be indistinguishable from a billing fault. Storage is a
level, like a seat count.

The weight does not fit in the column and the numbers are not close.
CeilingHolding.Held was an int: int.MaxValue is 2,147,483,647 and a 200GB
allowance is 214,748,364,800, a hundred times larger — even a 3GB world
overflows it. Held and the new CeilingClaim.Weight are long, the allowance and
every comparison widen with them, and the overflow test lives in the LocalDB
lane because a column-width truncation is invisible to a provider with no
columns. A wrapped aggregate does not fail loudly; it under-counts, so the
ceiling stops refusing and the bill is found later.

Quota moves when ownership moves, which was a hole between two individually
correct features. Transfer left the former owner's claims held indefinitely —
paying for storage they no longer had — while the recipient got it for free with
no capacity check. Transfer is now a claim migration: claim against the
recipient FIRST, release the sender second. Never the reverse, because releasing
first opens a window where a concurrent claim consumes the freed allowance and
the transfer then fails with the original owner's claim already gone. The
billing subscriber is stored on the workspace and is deliberately separate from
OwnerPartyId, because an Arke Kids workspace owned by a parent may be billed to
a family subscription.

And the elegance of keying the claim on the content hash is exactly what makes
release ambiguous. Idempotency per hash means a subscriber has ONE claim
covering every workspace of theirs that references it, so transferring one of
two such workspaces and releasing "their claim for that hash" would leave the
retained workspace's bytes completely uncharged while the physical blob cannot
be swept and we are still paying for it. BlobPossession is the per-subscriber
reference count that fixes it — the same shape as WorkspaceBlob.RefCount one
level up, answering a genuinely different question: physical RefCount answers
"may these bytes be deleted?", this answers "may this subscriber stop paying for
them?"

Registering `workspace` as a ShareResourceKind inherits every Spec 086 mechanic
with no new code, and the resolver returns only workspaces the owner actually
owns — the omission is load-bearing, because grant creation detects a caller
naming someone else's ids by comparing counts.

Migration is CLI-generated: Held widened to bigint, Weight and AccessLevel
added, BillingSubscriber* on workspaces, and AnkBlobPossessions with a filtered
unique index. Verified by applying the whole stream to a fresh LocalDB and
reading back Held = bigint.

Full solution suite green (2206 Application.Tests, 99 Database, 57 Architecture,
306 Api).

+++ src/Aonik.Subscriptions/Entities/Usage/CeilingHolding.cs
@@ -20,8 +20,17 @@ public class CeilingHolding : AuditableEntity, ITenantScoped
 
     public string MeterCode { get; set; } = string.Empty;
 
-    /// <summary>Slots currently claimed.</summary>
-    public int Held { get; set; }
+    /// <summary>
+    /// Slots currently claimed — <strong><c>long</c>, not <c>int</c></strong> (Spec 089 §9.1).
+    ///
+    /// <para>
+    /// Every ceiling to date has counted seats and profiles, where <c>int</c> was ample. A byte ceiling is not
+    /// close: <c>int.MaxValue</c> is 2,147,483,647 and a 200GB allowance is 214,748,364,800 — exactly one
+    /// hundred times larger, and even a 3GB world overflows it. A wrapped aggregate does not fail loudly; it
+    /// under-counts, so the ceiling stops refusing and the storage bill is discovered later.
+    /// </para>
+    /// </summary>
+    public long Held { get; set; }
 
     public List<CeilingClaim> Claims { get; set; } = new();
 }
